requirements.txt
flask
requests
beautifulsoup4
gunicorn
psycopg2-binary